+++ db/migrate/20260921200000_rename_fine_print_tables_to_agreement_engine.rb
@@ -1,0 +1,5 @@
+class RenameFinePrintTablesToAgreementEngine < ActiveRecord::Migration[8.0]
+  def change
+    rename_table :fine_print_documents, :agreement_engine_documents
+  end
+end