+++ test/integration/agreement_flow_test.rb
@@ -2,13 +2,13 @@ require "test_helper"
 
 class AgreementFlowTest < ActionDispatch::IntegrationTest
   setup do
-    @terms_v1 = FinePrint::Document.create!(
+    @terms_v1 = AgreementEngine::Document.create!(
       document_type: "terms_of_service",
       version: "1.0",
       content: "Terms v1 content",
       effective_at: 2.days.ago
     )
-    @privacy_v1 = FinePrint::Document.create!(
+    @privacy_v1 = AgreementEngine::Document.create!(
       document_type: "privacy_policy",
       version: "1.0",
       content: "Privacy v1 content",
@@ -36,7 +36,7 @@ class AgreementFlowTest < ActionDispatch::IntegrationTest
     sign_in @user
 
     get root_path
-    assert_redirected_to fine_print.agreement_path(:terms_of_service)
+    assert_redirected_to agreement_engine.agreement_path(:terms_of_service)
   end
 
   test "accepting agreement redirects back" do
@@ -46,7 +46,7 @@ class AgreementFlowTest < ActionDispatch::IntegrationTest
     )
     sign_in @user
 
-    patch fine_print.agreement_path(:terms_of_service)
+    patch agreement_engine.agreement_path(:terms_of_service)
     assert_response :see_other
 
     @user.reload
@@ -60,7 +60,7 @@ class AgreementFlowTest < ActionDispatch::IntegrationTest
     )
     sign_in @user
 
-    delete fine_print.agreement_path(:terms_of_service)
+    delete agreement_engine.agreement_path(:terms_of_service)
     assert_response :see_other
     assert_match(/please sign in and accept/i, flash[:alert])
   end